requirements.txt
streamlit>=1.32.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
openai>=1.30.0
rouge-score>=0.1.2
scikit-learn>=1.4.0
nltk>=3.8.0
psutil>=5.9.0
python-dotenv>=1.0.0
